Code deploy

+++ scripts/stop_server.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
 
-echo "Stop Server...."
-kill $(lsof -t -i:7998)
+if lsof -Pi :7998 -sTCP:LISTEN -t >/dev/null ; then
+        echo "Stopping Server...."
+        kill $(lsof -t -i:7998)
+else
+        echo "no server found"
+fi
